// src/app/Models/storeOwnerForAdd.ts
export interface StoreOwnerForAdd {
        email: string;
        firstName: string;
        lastName: string;
        mobile: number;
        storeMobile_1: number;
        storeMobile_2: number;
        street: string;
        city: string;
        government: string;
        storeName: string;
        whatsApp: number;
        facebookLink: string;
        password: string;
        confirmPassword: string;
}